+++ Assignment-3/src/middlewares/queryValidaton.js
@@ -3,11 +3,9 @@ const Joi = require("joi");
 const queryValidation = (req, res, next) => {
     const { id } = req.params;
     const {error} = Joi.number().validate(id)
-    console.log(error)
-    // const { error } = querySchema.validate({id});
-    // if (error) {
-    //     next(error);
-    // }
-    // next();
+    if (error) {
+        next(error);
+    }
+    next();
 }
 module.exports = queryValidation
